reorder chess module for better scalability
