+++ src-tauri/src/chess/pieces.rs
@@ -1,3 +1,5 @@
+//! Chess pieces logic
+
 #[derive(Debug)] // include this line right before your struct definition
 pub struct Piece {
     square: (char, u8),
